+++ lib/src/widget/attach_dialog_widget.dart
@@ -296,8 +296,9 @@ class _AttachDialogWidgetState extends State<AttachDialogWidget>
 
   ///等待动画结束,关闭动画资源
   Future<void> dismiss() async {
+    if (_ctrlBg == null) return;
     //over animation
-    _ctrlBg?.reverse();
+    _ctrlBg!.reverse();
     _ctrlBody.reverse();
 
     if (widget.useAnimation) {
@@ -308,6 +309,7 @@ class _AttachDialogWidgetState extends State<AttachDialogWidget>
   @override
   void dispose() {
     _ctrlBg?.dispose();
+    _ctrlBg = null;
     _ctrlBody.dispose();
     super.dispose();
   }
@@ -322,7 +324,6 @@ class AttachDialogController extends BaseController {
 
   @override
   Future<void> dismiss() async {
-    await Future.delayed(Duration(milliseconds: 1));
     try {
       await _state?.dismiss();
     } catch (e) {